requirements.txt
torch>=1.12.0
numpy>=1.21.0
scipy>=1.7.0
pandas>=1.3.0
scikit-learn>=1.0.0
tqdm>=4.62.0
pyyaml>=6.0
requests>=2.26.0
gdown>=4.5.0
matplotlib>=3.5.0
seaborn>=0.11.0
networkx>=2.6.0
tensorboard>=2.10.0